python_p/requirements.txt
fastapi
grpcio
grpcio-tools
uvicorn
requests
